+++ apps/wasm/mvp/.eslintrc.js
@@ -9,6 +9,6 @@ module.exports = {
   ],
   rules: {
     // override/add rules settings here, such as:
-    // 'vue/no-unused-vars': 'error'
+    'vue/multi-word-component-names': 'off'
   }
 }
