+++ apps/wasm/mvp/.parcelrc
@@ -1,7 +1,15 @@
 {
     "extends": "@parcel/config-default",
     "transformers": {
-        "*.wasm": ["@parcel/transformer-raw"]
+        "*.wasm": [
+            "@parcel/transformer-raw"
+        ],
+        "*.md": [
+            "@parcel/transformer-inline-string"
+        ]
     },
-    "reporters":  ["...", "parcel-reporter-static-files-copy"]
+    "reporters": [
+        "...",
+        "parcel-reporter-static-files-copy"
+    ]
 }
