+++ apps/wasm/mvp/package.json
@@ -5,7 +5,7 @@
   "private": true,
   "browserslist": "> 0.5%, last 2 versions, not dead",
   "scripts": {
-    "start": "parcel serve web/index.html --open --lazy",
+    "start": "parcel serve web/*.html --open",
     "watch:ext": "parcel watch chrome-ext/{background.ts,content_script.ts,*.html} --no-hmr",
     "patch-wasm": "node patch.js",
     "build:proto": "rimraf web/pb && node protogen.js",
@@ -26,31 +26,38 @@
   "dependencies": {
     "events": "^3.3.0",
     "lodash": "^4.17.21",
+    "marked": "^4.0.13",
     "protobufjs": "^6.11.2",
     "vue": "^3.2.31"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.8",
-    "@parcel/transformer-raw": "^2.4.0",
-    "@parcel/transformer-sass": "^2.4.0",
-    "@parcel/transformer-vue": "^2.4.0",
+    "@babel/core": "^7.17.9",
+    "@parcel/transformer-inline-string": "^2.4.1",
+    "@parcel/transformer-raw": "^2.4.1",
+    "@parcel/transformer-sass": "^2.4.1",
+    "@parcel/transformer-vue": "^2.4.1",
+    "@tailwindcss/forms": "^0.5.0",
+    "@tailwindcss/typography": "^0.5.2",
     "@types/chrome": "^0.0.180",
     "@types/jest": "^27.4.1",
-    "@types/lodash": "^4.14.180",
+    "@types/lodash": "^4.14.181",
     "@types/node": "latest",
+    "autoprefixer": "^10.4.4",
     "eslint": "^8.12.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^8.5.0",
+    "eslint-plugin-vue": "^8.6.0",
     "jest": "^27.5.1",
-    "parcel": "^2.4.0",
+    "parcel": "^2.4.1",
     "parcel-reporter-static-files-copy": "^1.3.4",
     "path-browserify": "^1.0.1",
+    "postcss": "^8.4.12",
     "postcss-html": "^1.3.0",
     "process": "^0.11.10",
     "rimraf": "^3.0.2",
-    "sass": "^1.49.9",
+    "sass": "^1.50.0",
     "stylelint": "^14.6.1",
     "stylelint-config-recommended-vue": "^1.4.0",
+    "tailwindcss": "^3.0.23",
     "ts-jest": "^27.1.4",
     "typescript": "^4.6.3",
     "util": "^0.12.4"
