+++ apps/wasm/mvp/src/js_exports.rs
@@ -1,9 +1,6 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0
 
-// https://github.com/rustwasm/wasm-bindgen/issues/2774#issuecomment-1030747023
-#![allow(clippy::unused_unit)]
-
 use crate::*;
 use js_sys::Uint8Array;
 use wasm_bindgen::prelude::*;
