--- apps/wasm/mvp/web/App.vue
@@ -1,133 +1,0 @@
-<template>
-  <div><h2>Mina Web Node Demo</h2></div>
-  <p>
-    Run
-    <a href="https://github.com/ChainSafe/mina-rs/tree/mvp-proxy/apps/mvp-proxy">proxy node</a>
-    locally, then paste node address into below input box.
-  </p>
-  <p />
-  <p>
-    Local node address: <br>
-    <input
-      v-model="addr"
-      class="addr"
-    > <br>
-    <button
-      v-if="wasmLoaded"
-      @click="connect"
-    >
-      Connect
-    </button>
-    <button
-      v-if="!wasmLoaded"
-      disabled="disabled"
-      @click="connect"
-    >
-      Loading
-    </button>
-  </p>
-  <div>
-    Peers({{ getConnectedPeerCount() }} connected / {{ getPeerCount() }} total):
-    <table>
-      <tr>
-        <th>id</th>
-        <th>connected</th>
-        <th>sync_status</th>
-        <th>protocol_state_hash</th>
-        <th>git_commit</th>
-        <th>uptime_minutes</th>
-      </tr>
-      <tr
-        v-for="id in peerKeys"
-        :key="id"
-      >
-        <td>{{ id }}</td>
-        <td>{{ peers[id].connected }}</td>
-        <td>{{ peers[id].syncStatus }}</td>
-        <td>{{ peers[id].protocolStateHash }}</td>
-        <td>{{ peers[id].gitCommit }}</td>
-        <td>{{ peers[id].uptimeMinutes }}</td>
-      </tr>
-    </table>
-  </div>
-</template>
-
-<script lang="ts">
-import { initWasm } from "./../utils";
-import { connect, set_event_emitter } from "./../pkg/wasm";
-import { EventEmitter } from "events";
-import { ConnectRequest } from "./pb/requests";
-import { PeerStatus } from "./pb/messages";
-import _ from "lodash";
-
-export default {
-  data() {
-    return {
-      addr: "/ip4/127.0.0.1/tcp/23333/ws/p2p/12D3KooWFxXTHUv5Cpiq5gFJ8JUWW1LAwxcGJdaHt4S4GFZU8PgC",
-      eventEmitter: new EventEmitter(),
-      peers: {},
-      peerKeys: [],
-      wasmLoaded: false,
-    };
-  },
-  created() {
-    this.loadWasm();
-  },
-  mounted() {
-    this.eventEmitter.on("log", (msg) => {
-      console.log(`[log] ${msg}`);
-    });
-    this.eventEmitter.on("update", (msg) => {
-      console.log(`[update] raw msg: ${msg}`);
-      try {
-        let ps = PeerStatus.decode(msg);
-        console.log(`[update] decoded msg: ${ps}`);
-        this.peers[ps.peerId] = ps;
-        this.peerKeys = Object.keys(this.peers);
-      } catch (e) {
-        console.log(e);
-        console.log(msg);
-      }
-    });
-  },
-  methods: {
-    async loadWasm() {
-      await initWasm();
-      this.wasmLoaded = true;
-    },
-    async connect() {
-      if (!this.wasmLoaded) {
-        return;
-      }
-      this.clear();
-      set_event_emitter(this.eventEmitter);
-      console.log(`[JS] Connecting to ${this.addr}`);
-      let req = ConnectRequest.create();
-      req.address = this.addr;
-      try {
-        await connect(ConnectRequest.encode(req).finish());
-      } catch (e) {
-        alert(e);
-      }
-    },
-    clear() {
-      this.peers = {};
-      this.peerKeys = [];
-    },
-    getPeerCount() {
-      return this.peerKeys.length;
-    },
-    getConnectedPeerCount() {
-      return _.chain(this.peers)
-        .filter((p) => p.connected)
-        .size()
-        .value();
-    },
-  },
-};
-</script>
-<style lang="scss" scoped>
-input.addr {
-  width: 800px;
-}
-</style>
