+++ apps/wasm/mvp/web/about.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { createApp } from "vue";
-import App from "./pages/index.vue";
+import App from "./pages/about.vue";
 
 const app = createApp(App);
 app.mount("#app");
