+++ apps/wasm/mvp/web/libp2p.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>mina web node</title>
+    <meta charset="utf-8" />
+    <link rel="stylesheet" href="styles.scss" />
+</head>
+
+<body>
+    <div id="app"></div>
+    <script type="module" src="libp2p.ts"></script>
+</body>
+
+</html>
