+++ apps/wasm/mvp/web/libp2p.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { createApp } from "vue";
-import App from "./pages/index.vue";
+import App from "./pages/libp2p.vue";
 
 const app = createApp(App);
 app.mount("#app");
