+++ apps/wasm/mvp/web/pages/about.vue
@@ -1,0 +1,17 @@
+<template>
+  <NavBar />
+  <div class="container px-4 prose">
+    <h2>About</h2>
+  </div>
+</template>
+
+<script lang="ts">
+import NavBar from "~/web/components/NavBar.vue";
+
+export default {
+  components: {
+    NavBar,
+  },
+};
+</script>
+<style lang="scss" scoped></style>
