+++ apps/wasm/mvp/web/styles.scss
@@ -1,1 +1,3 @@
-body{}
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
